@@ -2,14 +2,15 @@
   "name": "git-timecard",
   "description": "Single-user git based timecard.",
   "version": "0.6.0",
+  "type": "module",
   "bin": {
-    "gtc": "cli.mjs"
+    "gtc": "cli.js"
   },
   "scripts": {
-    "test": "healthier && node cli.mjs",
+    "test": "healthier && node cli.js",
     "posttest": "npm run format",
     "prepare": "husky install",
-    "format": "prettier --loglevel warn --write \"**/*.{json,js,mjs,css,md}\"",
+    "format": "prettier --loglevel warn --write \"**/*.{json,js,css,md}\"",
     "fix": "verdaccio-fix"
   },
   "repository": {
@@ -22,14 +23,14 @@
     "url": "http://robin.millette.info"
   },
   "engines": {
-    "node": ">= 14.0.0"
+    "node": ">= 14.17.0"
   },
   "license": "AGPL-3.0",
   "lint-staged": {
     "*.{js,json,md}": [
       "prettier --write"
     ],
-    "*.{json,js,mjs,css,md}": "prettier --write"
+    "*.{json,js,css,md}": "prettier --write"
   },
   "dependencies": {
     "isomorphic-git": "^1.10.1",
